optimize bytedance adapter
