+++ cocos/physics/physx/export-physx.ts
@@ -207,3 +207,168 @@ export function getShapeMaterials (pxMtl: any) {
     }
 }
 
+export function setupCommonCookingParam (params: any, skipMeshClean = false, skipEdgedata = false): void {
+    params.setSuppressTriangleMeshRemapTable(true);
+    if (!skipMeshClean) {
+        params.setMeshPreprocessParams(params.getMeshPreprocessParams() & ~PX.MeshPreprocessingFlag.eDISABLE_CLEAN_MESH);
+    } else {
+        params.setMeshPreprocessParams(params.getMeshPreprocessParams() | PX.MeshPreprocessingFlag.eDISABLE_CLEAN_MESH);
+    }
+
+    if (skipEdgedata) {
+        params.setMeshPreprocessParams(params.getMeshPreprocessParams() & ~PX.MeshPreprocessingFlag.eDISABLE_ACTIVE_EDGES_PRECOMPUTE);
+    } else {
+        params.setMeshPreprocessParams(params.getMeshPreprocessParams() | PX.MeshPreprocessingFlag.eDISABLE_ACTIVE_EDGES_PRECOMPUTE);
+    }
+}
+
+export function createConvexMesh (vertices: Float32Array, cooking: any, physics: any): any {
+    if (USE_BYTEDANCE) {
+        const cdesc = new PX.ConvexMeshDesc();
+        cdesc.setPointsData(vertices);
+        cdesc.setPointsCount(vertices.length / 3);
+        cdesc.setPointsStride(3 * Float32Array.BYTES_PER_ELEMENT);
+        cdesc.setConvexFlags(PX.ConvexFlag.eCOMPUTE_CONVEX);
+        return cooking.createConvexMesh(cdesc);
+    } else {
+        ;
+        const l = vertices.length;
+        const vArr = new PX.PxVec3Vector();
+        for (let i = 0; i < l; i += 3) {
+            vArr.push_back({ x: vertices[i], y: vertices[i + 1], z: vertices[i + 2] });
+        }
+        return cooking.createConvexMesh(vArr, physics);
+    }
+}
+
+//eTIGHT_BOUNDS = (1<<0) convex
+//eDOUBLE_SIDED = (1<<1) trimesh
+export function createMeshGeometryFlags (flags: number, isConvex: boolean) {
+    if (USE_BYTEDANCE) {
+        return flags;
+    } else {
+        return isConvex ? new PX.PxConvexMeshGeometryFlags(flags) : new PX.PxMeshGeometryFlags(flags);
+    }
+}
+
+export function createTriangleMesh (vertices: Float32Array, indices: Uint32Array, cooking: any, physics: any): any {
+    if (USE_BYTEDANCE) {
+        const meshDesc = new PX.TriangleMeshDesc();
+        meshDesc.setPointsData(vertices);
+        meshDesc.setPointsCount(vertices.length / 3);
+        meshDesc.setPointsStride(Float32Array.BYTES_PER_ELEMENT * 3);
+        meshDesc.setTrianglesData(indices);
+        meshDesc.setTrianglesCount(indices.length / 3);
+        meshDesc.setTrianglesStride(Uint32Array.BYTES_PER_ELEMENT * 3);
+        return cooking.createTriangleMesh(meshDesc);
+    } else {
+        const l = vertices.length;
+        const l2 = indices.length;
+        const vArr = new PX.PxVec3Vector();
+        for (let i = 0; i < l; i += 3) {
+            vArr.push_back({ x: vertices[i], y: vertices[i + 1], z: vertices[i + 2] });
+        }
+        const iArr = new PX.PxU16Vector();
+        for (let i = 0; i < l2; i += 3) {
+            iArr.push_back(indices[i]); iArr.push_back(indices[i + 1]); iArr.push_back(indices[i + 2]);
+        }
+        return cooking.createTriMeshExt(vArr, iArr, physics);
+    }
+}
+
+export function createBV33TriangleMesh (vertices: Float32Array, indices: Uint32Array, cooking: any, physics: any,
+    skipMeshCleanUp = false,
+    skipEdgeData = false,
+    cookingPerformance = false,
+    meshSizePerfTradeoff = true, inserted = true): any {
+    if (!USE_BYTEDANCE) return;
+    const meshDesc = new PX.TriangleMeshDesc();
+    meshDesc.setPointsData(vertices);
+    meshDesc.setPointsCount(vertices.length / 3);
+    meshDesc.setPointsStride(Float32Array.BYTES_PER_ELEMENT * 3);
+    meshDesc.setTrianglesData(indices);
+    meshDesc.setTrianglesCount(indices.length / 3);
+    meshDesc.setTrianglesStride(Uint32Array.BYTES_PER_ELEMENT * 3);
+
+    const params = cooking.getParams();
+    setupCommonCookingParam(params, skipMeshCleanUp, skipEdgeData);
+    const midDesc = new PX.BVH33MidphaseDesc();
+
+    if (cookingPerformance) midDesc.setMeshCookingHint(PX.MeshCookingHint.eCOOKING_PERFORMANCE);
+    else midDesc.setMeshCookingHint(PX.MeshCookingHint.eSIM_PERFORMANCE);
+
+    if (meshSizePerfTradeoff) midDesc.setMeshSizePerformanceTradeOff(0.0);
+    else midDesc.setMeshSizePerformanceTradeOff(0.55);
+
+    params.setMidphaseDesc(midDesc);
+    cooking.setParams(params);
+    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+    console.log(`Cook 状态：${cooking.validateTriangleMesh(meshDesc)}`);
+    return cooking.createTriangleMesh(meshDesc);
+}
+
+export function createBV34TriangleMesh (vertices: Float32Array, indices: Uint32Array, cooking: any, physics: any,
+    skipMeshCleanUp = false,
+    skipEdgeData = false,
+    numTrisPerLeaf = true,
+    inserted = true): void {
+    if (!USE_BYTEDANCE) return;
+    const meshDesc = new PX.TriangleMeshDesc();
+    meshDesc.setPointsData(vertices);
+    meshDesc.setPointsCount(vertices.length / 3);
+    meshDesc.setPointsStride(Float32Array.BYTES_PER_ELEMENT * 3);
+    meshDesc.setTrianglesData(indices);
+    meshDesc.setTrianglesCount(indices.length / 3);
+    meshDesc.setTrianglesStride(Uint32Array.BYTES_PER_ELEMENT * 3);
+    const params = cooking.getParams();
+    setupCommonCookingParam(params, skipMeshCleanUp, skipEdgeData);
+
+    const midDesc = new PX.BVH34MidphaseDesc();
+    midDesc.setNumPrimsLeaf(numTrisPerLeaf);
+    params.setMidphaseDesc(midDesc);
+    cooking.setParams(params);
+    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+    console.log(`Cook 状态：${cooking.validateTriangleMesh(meshDesc)}`);
+    return cooking.createTriangleMesh(meshDesc);
+}
+
+export function createHeightField (terrain: any, heightScale: number, cooking: any, physics: any) {
+    const sizeI = terrain.getVertexCountI();
+    const sizeJ = terrain.getVertexCountJ();
+    if (USE_BYTEDANCE) {
+        const samples = new PX.HeightFieldSamples(sizeI * sizeJ);
+        for (let i = 0; i < sizeI; i++) {
+            for (let j = 0; j < sizeJ; j++) {
+                const s = terrain.getHeight(i, j) / heightScale;
+                const index = j + i * sizeJ;
+                samples.setHeightAtIndex(index, s);
+                // samples.setMaterialIndex0AtIndex(index, 0);
+                // samples.setMaterialIndex1AtIndex(index, 0);
+            }
+        }
+        const hfdesc = new PX.HeightFieldDesc();
+        hfdesc.setNbRows(sizeJ);
+        hfdesc.setNbColumns(sizeI);
+        hfdesc.setSamples(samples);
+        return cooking.createHeightField(hfdesc);
+    } else {
+        const samples = new PX.PxHeightFieldSampleVector();
+        for (let i = 0; i < sizeI; i++) {
+            for (let j = 0; j < sizeJ; j++) {
+                const s = new PX.PxHeightFieldSample();
+                s.height = terrain.getHeight(i, j) / heightScale;
+                samples.push_back(s);
+            }
+        }
+        return cooking.createHeightFieldExt(sizeI, sizeJ, samples, physics);
+    }
+}
+
+export function createHeightFieldGeometry (hf: any, flags: number, hs: number, xs: number, zs: number) {
+    if (USE_BYTEDANCE) {
+        return new PX.HeightFieldGeometry(hf, hs, xs, zs);
+    } else {
+        return new PX.PxHeightFieldGeometry(hf, new PX.PxMeshGeometryFlags(flags),
+            hs, xs, zs);
+    }
+}
