+++ cocos/physics/physx/shapes/physx-box-shape.ts
@@ -1,6 +1,5 @@
 import { IVec3Like } from '../../../core';
-import { aabb, sphere } from '../../../core/geometry';
-import { Collider, RigidBody, PhysicMaterial, BoxCollider } from '../../framework';
+import { BoxCollider } from '../../framework';
 import { VEC3_0 } from '../../utils/util';
 import { IBoxShape } from '../../spec/i-physics-shape';
 import { PX } from '../export-physx';
