+++ cocos/physics/physx/shapes/physx-plane-shape.ts
@@ -1,6 +1,5 @@
 import { IVec3Like, Quat, Vec3 } from '../../../core';
-import { aabb, sphere } from '../../../core/geometry';
-import { Collider, RigidBody, PhysicMaterial, PlaneCollider } from '../../framework';
+import { PlaneCollider } from '../../framework';
 import { IPlaneShape } from '../../spec/i-physics-shape';
 import { getTempTransform, PX, _pxtrans, _trans } from '../export-physx';
 import { EPhysXShapeType, PhysXShape } from './physx-shape';
